add ff for opq in python/ & split linalg file into two modules
