+++ python/Cargo.toml
@@ -18,7 +18,7 @@ arrow-schema = "37.0"
 chrono = "0.4.23"
 env_logger = "0.10"
 futures = "0.3"
-lance = { path = "../rust"}
+lance = { path = "../rust" }
 log = "0.4"
 prost = "0.11"
 pyo3 = { version = "0.18.1", features = ["extension-module", "abi3-py38"] }
@@ -27,3 +27,6 @@ uuid = "1.3.0"
 
 [build-dependencies]
 prost-build = "0.11"
+
+[features]
+opq = ["lance/opq"]
