+++ python/src/dataset.rs
@@ -373,6 +373,12 @@ impl Dataset {
                     };
 
                     if let Some(o) = kwargs.get_item("use_opq") {
+                        #[cfg(not(feature = "opq"))]
+                        if PyAny::downcast::<PyBool>(o)?.extract()? {
+                            return Err(PyValueError::new_err(format!(
+                                "Feature 'opq' is not installed."
+                            )));
+                        }
                         pq_params.use_opq = PyAny::downcast::<PyBool>(o)?.extract()?
                     };
